+++ Sources/TuistKit/Commands/ScaffoldCommand.swift
@@ -29,6 +29,11 @@ struct ScaffoldCommand: ParsableCommand {
                              subcommands: [ListCommand.self])
     }
 
+    @Flag(
+        help: "The output in JSON format"
+    )
+    var json: Bool
+
     @Option(
         name: .shortAndLong,
         help: "The path to the folder where the template will be generated (Default: Current directory)"
@@ -49,6 +54,7 @@ struct ScaffoldCommand: ParsableCommand {
     init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         template = try container.decode(Argument<String>.self, forKey: .template).wrappedValue
+        json = try container.decodeIfPresent(Option<Bool>.self, forKey: .json)?.wrappedValue ?? false
         path = try container.decodeIfPresent(Option<String>.self, forKey: .path)?.wrappedValue
         try ScaffoldCommand.requiredTemplateOptions.forEach { option in
             requiredTemplateOptions[option.name] = try container.decode(Option<String>.self,
@@ -63,7 +69,8 @@ struct ScaffoldCommand: ParsableCommand {
     func run() throws {
         // Currently, @Argument and subcommand clashes, so we need to handle that ourselves
         if template == ListCommand.configuration.commandName {
-            try ListService().run(path: path)
+            let format: ListService.OutputFormat = json ? .json : .table
+            try ListService().run(path: path, outputFormat: format)
         } else {
             try ScaffoldService().run(path: path,
                                       templateName: template,
@@ -87,13 +94,13 @@ extension ScaffoldCommand {
         else { throw ScaffoldCommandError.templateNotProvided }
         guard !configuration.subcommands.contains(where: { $0.configuration.commandName == arguments[1] }) else { return }
         // We want to parse only the name of template, not its arguments which will be dynamically added
-        // Plucking out path argument
+        // Plucking out json and path arguments
         let pairedArguments: [[String]] = stride(from: 2, to: arguments.count, by: 2).map {
             Array(arguments[$0 ..< min($0 + 2, arguments.count)])
         }
         let filteredArguments = pairedArguments
             .filter {
-                $0.first == "--path" || $0.first == "-p"
+                $0.first == "--path" || $0.first == "-p" || $0.first == "--json"
             }
             .flatMap { $0 }
 
@@ -116,6 +123,7 @@ extension ScaffoldCommand {
 extension ScaffoldCommand {
     enum CodingKeys: CodingKey {
         case template
+        case json
         case path
         case required(String)
         case optional(String)
@@ -124,6 +132,8 @@ extension ScaffoldCommand {
             switch self {
             case .template:
                 return "template"
+            case .json:
+                return "json"
             case .path:
                 return "path"
             case let .required(required):
@@ -137,6 +147,8 @@ extension ScaffoldCommand {
             switch stringValue {
             case "template":
                 self = .template
+            case "json":
+                self = .json
             case "path":
                 self = .path
             default:
@@ -166,6 +178,7 @@ extension ScaffoldCommand: CustomReflectable {
             .map { Mirror.Child(label: $0.name, value: $0.option) }
         let children = [
             Mirror.Child(label: "template", value: _template),
+            Mirror.Child(label: "json", value: _json),
             Mirror.Child(label: "path", value: _path),
         ].filter {
             // Prefer attributes defined in a template if it clashes with predefined ones